refactor: extract styles to component and rename test component
