+++ app/components/UI/PermissionsSummary/PermissionsSummary.styles.ts
@@ -1,0 +1,85 @@
+import { StyleSheet } from 'react-native';
+import Device from '../../../util/device';
+import { Theme } from '../../../util/theme/models';
+
+const createStyles = (params: { theme: Theme }) => {
+  const { theme } = params;
+  return StyleSheet.create({
+    mainContainer: {
+      backgroundColor: theme.colors.background.default,
+      paddingTop: 24,
+      borderTopLeftRadius: 20,
+      borderTopRightRadius: 20,
+      minHeight: 200,
+      paddingBottom: Device.isIphoneX() ? 20 : 0,
+    },
+    title: {
+      fontSize: Device.isSmallDevice() ? 18 : 24,
+      marginBottom: 16,
+      marginTop: 16,
+      marginRight: 24,
+      marginLeft: 24,
+    },
+    actionButtonsContainer: {
+      flex: 0,
+      flexDirection: 'row',
+      padding: 24,
+    },
+    buttonPositioning: {
+      flex: 1,
+    },
+    cancelButton: {
+      marginRight: 8,
+    },
+    confirmButton: {
+      marginLeft: 8,
+    },
+    networkPermissionRequestInfoCard: {
+      marginHorizontal: 24,
+      marginTop: 8,
+      marginBottom: 12,
+      alignItems: 'center',
+      flexDirection: 'row',
+    },
+    permissionsSummaryHeader: {
+      justifyContent: 'center',
+      alignItems: 'center',
+    },
+    domainLogoContainer: {
+      width: 32,
+      height: 32,
+      borderRadius: 16,
+    },
+    assetLogoContainer: {
+      alignItems: 'center',
+      justifyContent: 'center',
+      borderRadius: 10,
+    },
+    networkPermissionRequestDetails: {
+      flex: 1,
+      marginLeft: 12,
+    },
+    permissionRequestNetworkInfo: {
+      flexDirection: 'row',
+      alignItems: 'center',
+    },
+    avatarGroup: { marginLeft: 2 },
+    accountPermissionRequestInfoCard: {
+      marginHorizontal: 24,
+      marginTop: 8,
+      marginBottom: 12,
+      alignItems: 'center',
+      flexDirection: 'row',
+    },
+    accountPermissionRequestDetails: {
+      flex: 1,
+      marginLeft: 12,
+    },
+    permissionRequestAccountInfo: {
+      flexDirection: 'row',
+      alignItems: 'center',
+    },
+  });
+};
+
+export default createStyles;
